+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Jobs can be now submitted from a remote machine.
 - Submitting jobs using `Submitter.submit` is now thread-safe.
 - Handling exceptions that occur when reading a configuration file.
+- When reading configuration file fails, an exception is no longer raised, instead an error is reported and a default configuration is used.
 
 ***